Move the huggingface package under modeldownloader
The huggingface package is only for model download.
